First pass on the cart form
- Adds .amp.erb versions of product page views
- adds a GET route to OrdersController#populate, and uses in _cart_form.amp.erb
- overrides Overrides ActionView::Helpers::FormTagHelper#form_tag_with_body to add _target attribute

This gets us successfully to populating the cart. However, it’s probably not the way Google intends things to be done. There are a lot of AMP validation errors on the product page, and I think a good chance that resolving them will be incompatible with this approach. But… the cart does populate.